feat(hud): Refactored the GameHUD into components since the file was getting a bit to big to navigate. Added fill effect when holding down the left click!
