+++ packages/frontend/src/game/game_hud/components/index.ts
@@ -1,0 +1,4 @@
+export { FistComponent } from './FistComponent';
+export { HpBarComponent } from './HpBarComponent';
+export { PlayerListComponent } from './PlayerListComponent';
+export { TimerComponent } from './TimerComponent';
